Updated adjacentMatrix from byte to int to handle weighted graph

+++ src/dsa/Graph.java
@@ -17,7 +17,7 @@ import java.io.PrintStream;
 import java.util.*;
 
 public class Graph {
-    public byte[][] adjacencyMatrix;
+    public int[][] adjacencyMatrix;
     public ArrayList<ArrayList<Integer>> adjacencyList;
 
     public Graph(int v){
@@ -25,10 +25,10 @@ public class Graph {
         for(int i = 1; i<=v; ++i){
             adjacencyList.add(new ArrayList<>());
         }
-        adjacencyMatrix = new byte[v][v];
+        adjacencyMatrix = new int[v][v];
     }
 
-    public Graph(int v, byte[][] matrix){
+    public Graph(int v, int[][] matrix){
         adjacencyList = new ArrayList<>();
         for(int i = 1; i<=v; ++i){
             adjacencyList.add(new ArrayList<>());
